requirements.txt
flask
flask_cors
flask_jwt_extended
bcrypt
psycopg2
python-dotenv
gunicorn
